feat: プレイ画面のロジックをフック化
